comit

+++ include/memory_region/mem_regs.h
@@ -5,13 +5,13 @@
 
 typedef struct s_mem_reg
 { 
-	u32 start_addr;
-	u32 end_addr;
+	u64 start_addr;
+	u64 end_addr;
 	
 } 
 t_mem_reg;
 
-t_mem_reg* create_mem_reg(u32 start_addr,u32 end_addr);
+t_mem_reg* create_mem_reg(u64 start_addr, u64 end_addr);
 void delete_mem_reg(t_mem_reg* mem_reg);
 
 #endif
