+++ memory_region/mem_regs.c
@@ -1,12 +1,12 @@
 #include "memory_region/mem_regs.h"
 
-t_mem_reg* create_mem_reg(u32 start_addr,u32 end_addr)
+t_mem_reg* create_mem_reg(u64 start_addr, u64 end_addr)
 {
-	t_mem_reg* mem_reg=NULL;
+	t_mem_reg* mem_reg = NULL;
 
-	mem_reg=kmalloc(sizeof(t_mem_reg));
-	mem_reg->start_addr=start_addr;
-	mem_reg->end_addr=end_addr;
+	mem_reg = kmalloc(sizeof(t_mem_reg));
+	mem_reg->start_addr = start_addr;
+	mem_reg->end_addr = end_addr;
 	return mem_reg;
 }
